Error minimized output
So for this one I ran the code again without Guam, USVI, Samoa, and the Northern Marianas in the pool, and with a base of 19 so that the lowest a state could get was 20.  This results in an output of over 10,000 members of the house of representatives and also narrows the final error of the the last state getting one  representative to below 5% in the direction of being over or under represented, which is quite interesting from a fairness perspective.